+++ src/user/user_init.c
@@ -243,6 +243,7 @@ void mjs_defaultMesh(mjsMesh* mesh) {
   memset(mesh, 0, sizeof(mjsMesh));
   mesh->refquat[0] = 1;
   mesh->scale[0] = mesh->scale[1] = mesh->scale[2] = 1;
+  mesh->maxhullvert = -1;
 }